Move many functions from superpmi.py to jitutil.py
Add downloading of portable git for use on Helix for jit-analyze.
It is added to the correlation payload that is copied to Helix machines
(currently, we only use Windows Helix machines).
